Added correct pin for input in targets.h. Added capture input and alternative pin function define variable for changing the input pin correctly. Also changed the CPU_FREQUENCY_MHZ to 96MHz as this is the CPU/System/Ctimer0 clock source.
removed old dshot STM init function from peripheral.c and peripherals.h.

Added unlock clock and freeze clock before PORT/GPIO/INPUTMUX initialization.
removed the enableComparator() function from the enablePeripherals().

Update Dshot DMA such that when it is set for receiving dshot/pwm the BITER_ELINKNO and the DLAST_SGA are also updated with the new buffersize value.
Also added that a dshot timing correction is only done when the buffersize is larger than 3.

Added the correct code in receiveDshotDma() for coniguring our capture timer and the dshot DMA. Also did some cleanup in the InputPin functions.

Added the correct code in the Dshot DMA interrupt handler so the Dshot/PWM is handled correctly.

Cleaned up timer.c init functions, so the modifyReg is used.
Correctly set the dshotPwmTimer initialization.

Fixed bug with Systick timer not working when starting the MCXA153 from a cold-boot without debugger. This problem is resolved by selecting the PE clock as the Systick clock source.