Update code (still unstable)
When controlling the motor control board with a PWM signal with 2.2ms period. The motor starts spinning from 50% duty cycle as expected but as the duty cycle reaches 55.7% the speed does not increase with the duty cycle.
By halving the interval_timer frequency to 1MHz, the motor speed is able to increase with the increasing duty cycle.

These changes have been made:
- Comparators are in normal mode. Hysteresis is disabled.
- The comparator output is being filtered inside the comparator unit to filter out the spikes that generate unwanted comparator triggers.
- Used the serial telemetry output for real-time debugging variables.
- Set the frequency of the interval_timer to 1MHz instead of 2MHz. This causes the motor to be able to spin to full speed.
- Disabled the fault protection in the PWM driver as this had no use.